+++ src/data_collection/processes/linux.rs
@@ -12,11 +12,10 @@ use hashbrown::HashSet;
 use process::*;
 use sysinfo::ProcessStatus;
 
-use super::{ProcessHarvest, UserTable};
+use super::{Pid, ProcessHarvest, UserTable};
 use crate::{
     data_collection::DataCollector,
     utils::error::{self, BottomError},
-    Pid,
 };
 
 /// Maximum character length of a /proc/<PID>/stat process name.
